Update patch set 2
Patch Set 2:

If we enable this feature, we will need a paired variable that crossystem can check that says "we just got reset/cleared" so that userspace knows more than just TPM is unowned, but also that it was a userspace triggered event and to _not_ trust any non-integrity checked data (no mounts, etc).

If not, then we should make this a recovery mode feature only and optionally add a crossystem  variable to be  checked ( to wipe stateful).  The signal of unowned TPM is fine though if phys presence was _definitely_ provided.  If not, we have to avoid leaving gaps.

Am I off in crazy-land or does this make sense?

Patch-set: 2