Change VbExEcGetExpectedRW to VbExEcGetRW and make vboot pick an ec image.
This change changes the function name as described above, and also changes its
signature so that it takes an additional parameter that specifies which EC
image vboot wants. This is better than making U-Boot decide because U-Boot
doesn't really keep track of which version it is (it peeks at internal vboot
data) and vboot does.

Also, some consts were removed from the image pointer pointer. The pointer
itself will be changed in the body of the function to tell vboot where the EC
has been loaded, and the contents of the buffer will be changed because U-Boot
will have to actually load the EC there.

BUG=chrome-os-partner:11148
TEST=Built vboot_reference, vboot_reference-firmware, chromeos-u-boot, and
chromeos-bootimage for Daisy and Link and saw them complete successfully with
and without the signature for U-Boot's version of this function being updated.
That works because the function isn't actually being used yet.

Change-Id: I2814c8210eb5b3d965bb8bbf23c0f283f9e44c90
Signed-off-by: Gabe Black <gabeblack@google.com>