Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

This change will make U-Boot break because it changes the signature of a function U-Boot has a stub definition for. To get it in, we'll have to first delete the stub from U-Boot, then commit this, and then reintroduce the stub or a real implementation in U-Boot. I'll make those changes once other pending related changes have made it into the repository.

Patch-set: 2
Label: Code-Review=-1
Label: Commit-Queue=0
Label: Verified=0