Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

(2 inline comments)

Something's not right. The Remove screens show both the yuck and the remove text, which is wrong (the inserted image might be the right one, but you still have to take it out first anyway).

Also the DEFAULT.yaml should be the auto-generated yaml that includes all the locales. We pare it down for the actual release.  This one cuts out ever screen (maybe a manual edit?)

Patch-set: 3
Label: Code-Review=-1
Label: Verified=0