Update patch set 4
Patch Set 4:

This CL is getting to be a bit of a monster.

I've updated the make_default_yaml script to allow for more flexible placement of images. Currently it runs slowly...about 1 second per language.

I've updated the localized_text for launch countries only. The new text was machine translated and will need fixing up before we use it. We also need text for the non-launch countries.

Patch-set: 4
Label: Code-Review=0
Label: Verified=0