Get kernel size/load address from vboot headers
Presently kernel load address and buffer size are programmed in the
u-boot device tree. There is no reason for this: the address and size
are part of the vboot encapsulation headers. Duplicating this
information hardcoded in the device tree does not bring any benefit
and is in fact harmful, as it is easy to get out of sync.

A better way of doing things is to derive kernel load address and size
from the appropriate vboot header. There is no sanity check yet, as it
is presumed that there is enough memory to load any kernel and u-boot
does not use the space above 0x100000 for at least 16 megabytes (the
kernel partition size). On x86 platform the check could be verify that
the top of the kernel space is well below the stack.

Instead of getting the kernel load information from u-boot, vboot now
derives the information from the headers and reports it back to u-boot
using the same structure.

BUG=chrome-os-partner:11994
TEST=manual
  . with the appropriate u-boot change run a Link target through a
    FAFT cycle, observe it succeed.

Change-Id: I3c2c2cefb1e31d16ac497a01894bf32638479ed7
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>