Update patch set 3
Patch Set 3:

If the firmware only indicated clear-done on the first boot and that flag auto-cleared, you could avoid the OS noticing simply by:
1) request clear
2) reboot
3) clear occurs
4) before system finishes booting, shut down system

So we need to keep the flag around until whatever cares about an owner-clear has had time to see it.

Whatever cares about that flag should clear it.

Patch-set: 3