Update patch set 2
Patch Set 2: (2 inline comments)

I've verified that the alignment is correct on a kiev.

The height of the hwid_placeholder.bmp matches the height of the HWID font characters because the default margin in build_font and text_to_png are the same...this is a little fragile... but it's better than using the hwid_unknown.bmp file to control the alignment.

Patch-set: 2