Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The audio loop is what provides the dev-screen timeout. VbAudioOpen() chooses a list of notes to play (which include silence) and when the list is exhausted the delay is over. 

There are two built-in lists: one for the normal 30-second delay, and another for the factory-mode 2-second delay.

If you short-circuit the list, the time delay goes away too and you'll never see the dev-screen or have a chance to interrupt it with a keystroke.

Patch-set: 1
Reviewer: Gerrit User 1000058 <1000058@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=0