Update patch set 1
Patch Set 1:

There is no capabiility to see a dev screen in the Sandbox U-boot at this time, nor is there an ability to play any sounds.  

The timeout, in sandbox mode only, only serves to delay the progress of testing for the timeout period.

Sandbox can run with the timeouts, but it's not really useful since all we care about (at this point) is *which* kernel & firmware have been selected.

Patch-set: 1