Write-protect the running RW BIOS before booting.
In normal or dev-mode, we want to write protect the RW BIOS that we're about
to use. This includes the case where we're staying in RO because the vblock
flags say to (we have to protect the vblock in that case, so we just go
ahead and protect the whole section).

We don't do if the flash isn't completely protecting the RO section, so that
we can continue to do full updates.

Of course, this depends on the BIOS passing in the correct flags and
implementing the protection scheme. Not all CPUs can do that.

BUG=chrome-os-partner:11215
BRANCH=link
TEST=manual

Make sure that the HW WP pin is asserted (crossystem wpsw_boot), and that
the flash SW protection bit has been set too (crossystem sw_wpsw_boot).

Boot normally, try to write the RW_SECTION_A:

  cd /tmp
  flashrom -r bios.bin
  dd if=/dev/zero of=foo bs=983040 count=1
  flashrom -w bios.bin -iRW_SECTION_A:foo

It should fail.

Writing RW_SECTION_B should work:

  flashrom -w bios.bin -iRW_SECTION_B:foo

Recovery mode should still work, of course.

If you disable the HW WP pin, then you should be able to trash both
RW_SECTION_A and RW_SECTION_B.

Change-Id: I97ea657047a6aeeff137925320a66fa12daf484a
Signed-off-by: Bill Richardson <wfrichar@chromium.org>