Update patch set 3
Patch Set 3:

We do not update firmware for Cr-48, so the more important cases to test are the Alex and ZGB.

The simplest way to check is to use the sign_official_build script that uses this to re-sign the firmware on an input image.

src/platform/vboot_reference/scripts/image_signing/sign_official_build.sh

You will probably want to have devices with dev-signed firmware on them. Then take an image from chromeos-images/, re-sign it using dev keys (which will also re-sign the firmware payload). And verify that the new image works (and succeeds with the firmware update).

Patch-set: 3
Label: Commit-Queue=0
Label: Verified=0