Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

As I reported on the bug, this change has a positive impact on idling the disk.  We need to decide if and how to change the commit time for the underlying ext4 mount, and we should probably make both changes at once.

Patch-set: 1