STOUT-FIRMWARE: Add the new GBB_FLAG_ENTER_TRIGGERS_TONORM usage
This GBB_FLAG_ENTER_TRIGGERS_TONORM is new added for allowing Enter
(default is Space) to trigger the TONORM screen. This flag is
used by FAFT testing.

Chage the usage of set_gbb_flags.sh to help an user.

Cherry-picked for faft support for Stout. This is not the complete
solution.

BRANCH=stout
BUG=chrome-os-partner:14056
TEST=None. Trivial comment change.

Change-Id: I8f734bd686beed761bdb8b5a2993cf32fec92bb4
Original-Change-Id: I8a1efa880627965570719f22dedfa0e31ab2d5f5
Signed-off-by: Marc Jones <marc.jones@chromium.org>