Update patch set 2
Patch Set 2: Verified

i talked to Don, and it doesn't look like anything uses the pubkey, so i dropped that logic.  this *almost* makes this target a generic hash-signer if it weren't for the hardcoded key ...

if we, in the future, want to sign hashes with other keys, we could extend insn format/signer further for that.  but for now, let's stick with this since we haven't had a need yet.

Patch-set: 2
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Verified=+1