BUTTERFLY-FIRMWARE: Print warning in TONORM when GBB flags force DEV.
We use gbb-flag-force-dev-switch-on in default firmware images to make
things easier for factory and some devs.

But when we request normal mode there should be some sort of warning/error
telling the user that this is not available, otherwise we can spent time trying
to figure out why the firmware doesn't work.

BRANCH=all
BUG=chrome-os-partner:14474
TEST=manual, test basic firmware functions.

Change-Id: I607b1a005632cd15bc7e109006e6095edfb20543
Original-Change-Id: Id48c12693c7575001fae7fad92a868cb5465e83d