Update patch set 1
Patch Set 1:

IMO, best-practices for a library are to never exit. It should always be up to the caller to decide how to handle errors. (Exceptions being absolutely unrecoverable situations, like glibc's heap memory corruption detection.)

Based on the plumbing of this library into firmware, it's not clear to me if there is a sensible way to handle error conditions -- I've not read that code, and so I was trying to shy away from changing the API too much.

Doing exit/error based on compile vs runtime seems wrong to me -- this just extends the behavior differences further. I think first steps should be to adopt this CL, and then move to eliminating the exit condition entirely. As an alternative, I could try removing the exit condition everywhere as part of this CL, too.

Patch-set: 1