vboot_next: Starting point for the vboot_next wrapper
Create a directory to work in, a do-nothing shim for the wrapper, an
extremely simple test, and modify the Makefiles to try it all when you run
"make vboot_next".

BUG=chromium-os:37062
BRANCH=none
TEST=none

It builds, but the tests all fail because it doesn't do anything yet.
There's no point in testing that.

Change-Id: I654c1444411e785869d5a5a88328f4df95f5f32a
Signed-off-by: Bill Richardson <wfrichar@chromium.org>