(modify and resend) Trigger sync right after writing.
This can trigger the sync mechanism earlier so that the kernel has enough
time to schedule a real disk sync before system reboots.

We sent this in CL 41269, however it is reverted because the sync() makes
test cases timeout (tons of disk access). So we only sync file itself in
this CL and leave the whoe system sync outside cgpt.

BUG=chromium-os:35992
TEST=compile only
BRANCH=none

Change-Id: I9c6b602220131f1f011e7865788a122d4c379e73